event: separate metric types with Record methods
This is an alternative to having metrics as builders. Instead we use
the same builder for everything, but BuilderMetric takes a type that
is a pair of a Metric and Value. These pairs are created by the Record
metric of each Metric type.

This has the main advantage of separate builders -- typed metric
values -- without the downsides of repeated code, both in the
implementation and the client.

Change-Id: I8d39880b364975416e6fe94f6545fd3fe201aa61
Reviewed-on: https://go-review.googlesource.com/c/exp/+/326390
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Ian Cottrell <iancottrell@google.com>